Moving DialogResponseMetadata to its rightful package
