+++ src/main/java/com/github/seratch/jslack/api/methods/response/dialog/DialogOpenResponse.java
@@ -1,7 +1,7 @@
 package com.github.seratch.jslack.api.methods.response.dialog;
 
 import com.github.seratch.jslack.api.methods.SlackApiResponse;
-import com.github.seratch.jslack.api.model.DialogResponseMetadata;
+import com.github.seratch.jslack.api.model.dialog.DialogResponseMetadata;
 import lombok.Data;
 
 @Data
