+++ src/main/java/com/github/seratch/jslack/api/model/dialog/DialogResponseMetadata.java
@@ -1,4 +1,4 @@
-package com.github.seratch.jslack.api.model;
+package com.github.seratch.jslack.api.model.dialog;
 
 import lombok.Data;
